+++ src/Converter/Data/c-d-overview.js
@@ -1,67 +1,67 @@
-import { timespanDiff } from '../../Utils/TimeHelper'
+// import { timespanDiff } from '../../Utils/TimeHelper'
 
-const convert = ({
-  collection,
-  timespan,
-  startDate,
-  endDate,
-  workingDate,
-  nonWorkingDateCount,
-}) => {
-  let pv = 0,
-    uv = 0,
-    buyerCount = 0,
-    supplierCount = 0
+// const convert = ({
+//   collection,
+//   timespan,
+//   startDate,
+//   endDate,
+//   workingDate,
+//   nonWorkingDateCount,
+// }) => {
+//   let pv = 0,
+//     uv = 0,
+//     buyerCount = 0,
+//     supplierCount = 0
 
-  const source = [],
-    country = [],
-    sourceObj = {},
-    countryObj = {},
-    idxOffset = workingDate === true ? 1 : 0,
-    channelIdx = 0 + idxOffset,
-    countryIdx = 1 + idxOffset,
-    pvIdx = 2 + idxOffset,
-    uvIdx = 3 + idxOffset,
-    buyerCountIdx = 4 + idxOffset,
-    supplierCountIdx = 5 + idxOffset
+//   const source = [],
+//     country = [],
+//     sourceObj = {},
+//     countryObj = {},
+//     idxOffset = workingDate === true ? 1 : 0,
+//     channelIdx = 0 + idxOffset,
+//     countryIdx = 1 + idxOffset,
+//     pvIdx = 2 + idxOffset,
+//     uvIdx = 3 + idxOffset,
+//     buyerCountIdx = 4 + idxOffset,
+//     supplierCountIdx = 5 + idxOffset
 
-  let days = timespanDiff(timespan || 30, startDate, endDate)
+//   let days = timespanDiff(timespan || 30, startDate, endDate)
 
-  days = workingDate === true ? days - nonWorkingDateCount : days
+//   days = workingDate === true ? days - nonWorkingDateCount : days
 
-  collection.forEach(item => {
-    pv += parseInt(item[pvIdx], 10)
-    uv += parseInt(item[uvIdx], 10)
-    buyerCount += parseInt(item[buyerCountIdx])
-    supplierCount += parseInt(item[supplierCountIdx])
-    if (!(item[channelIdx] in sourceObj)) {
-      sourceObj[item[channelIdx]] = true
-    }
-    if (!(item[countryIdx] in countryObj)) {
-      countryObj[item[countryIdx]] = true
-    }
-  })
+//   collection.forEach(item => {
+//     pv += parseInt(item[pvIdx], 10)
+//     uv += parseInt(item[uvIdx], 10)
+//     buyerCount += parseInt(item[buyerCountIdx])
+//     supplierCount += parseInt(item[supplierCountIdx])
+//     if (!(item[channelIdx] in sourceObj)) {
+//       sourceObj[item[channelIdx]] = true
+//     }
+//     if (!(item[countryIdx] in countryObj)) {
+//       countryObj[item[countryIdx]] = true
+//     }
+//   })
 
-  pv = Math.round(pv / days)
-  uv = Math.round(uv / days)
+//   pv = Math.round(pv / days)
+//   uv = Math.round(uv / days)
 
-  // convert source
-  Object.keys(sourceObj).forEach(item => {
-    source.push({
-      text: item,
-      value: item,
-    })
-  })
+//   // convert source
+//   Object.keys(sourceObj).forEach(item => {
+//     source.push({
+//       text: item,
+//       value: item,
+//     })
+//   })
 
-  // convert country
-  Object.keys(countryObj).forEach(item => {
-    country.push({
-      text: item,
-      value: item,
-    })
-  })
+//   // convert country
+//   Object.keys(countryObj).forEach(item => {
+//     country.push({
+//       text: item,
+//       value: item,
+//     })
+//   })
 
-  return { pv, uv, buyerCount, supplierCount, source, country }
-}
+//   return { pv, uv, buyerCount, supplierCount, source, country }
+// }
 
-export default convert
+// export default convert
