+++ src/assets/client.js
@@ -8,33 +8,33 @@ okchart.init({
       type: 'overview',
       container: 'overview-container',
     },
-    {
-      type: 'pv-uv',
-      container: 'pv-uv-container',
-    },
-    {
-      type: 'distribution',
-      container: 'distribution-container',
-    },
-    {
-      type: 'source-top-15',
-      container: 'source-top-15-container',
-    },
-    {
-      type: 'buyers-regist',
-      container: 'buyers-regist-container',
-    },
-    {
-      type: 'buyers-regist-distribution',
-      container: 'buyers-regist-distribution-container',
-    },
-    {
-      type: 'suppliers-regist',
-      container: 'suppliers-regist-container',
-    },
-    {
-      type: 'suppliers-regist-distribution',
-      container: 'suppliers-regist-distribution-container',
-    },
+    // {
+    //   type: 'pv-uv',
+    //   container: 'pv-uv-container',
+    // },
+    // {
+    //   type: 'distribution',
+    //   container: 'distribution-container',
+    // },
+    // {
+    //   type: 'source-top-15',
+    //   container: 'source-top-15-container',
+    // },
+    // {
+    //   type: 'buyers-regist',
+    //   container: 'buyers-regist-container',
+    // },
+    // {
+    //   type: 'buyers-regist-distribution',
+    //   container: 'buyers-regist-distribution-container',
+    // },
+    // {
+    //   type: 'suppliers-regist',
+    //   container: 'suppliers-regist-container',
+    // },
+    // {
+    //   type: 'suppliers-regist-distribution',
+    //   container: 'suppliers-regist-distribution-container',
+    // },
   ],
 })
